requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
requests
python-dotenv
pytest